// src/generated/rustplus.ts
export interface AppTeamInfo {
  leaderSteamId: string;
  members: AppTeamInfo_Member[];
  mapNotes: AppTeamInfo_Note[];
  leaderMapNotes: AppTeamInfo_Note[];
}

export interface AppTeamInfo_Member {
  steamId: string;
  name: string;
  x: number;
  y: number;
  isOnline: boolean;
  spawnTime: number;
  isAlive: boolean;
  deathTime: number;
}

export interface AppTeamInfo_Note {
  type: number;
  x: number;
  y: number;
}
// ...
export const AppTeamInfo: MessageFns<AppTeamInfo> = {
  encode(message: AppTeamInfo, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.leaderSteamId !== "0") {
      writer.uint32(8).uint64(message.leaderSteamId);
    }
    for (const v of message.members) {
      AppTeamInfo_Member.encode(v!, writer.uint32(18).fork()).join();
    }
    for (const v of message.mapNotes) {
      AppTeamInfo_Note.encode(v!, writer.uint32(26).fork()).join();
    }
    for (const v of message.leaderMapNotes) {
      AppTeamInfo_Note.encode(v!, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AppTeamInfo {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAppTeamInfo();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.leaderSteamId = reader.uint64().toString();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.members.push(AppTeamInfo_Member.decode(reader, reader.uint32()));
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.mapNotes.push(AppTeamInfo_Note.decode(reader, reader.uint32()));
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.leaderMapNotes.push(AppTeamInfo_Note.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<AppTeamInfo>, I>>(base?: I): AppTeamInfo {
    return AppTeamInfo.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AppTeamInfo>, I>>(object: I): AppTeamInfo {
    const message = createBaseAppTeamInfo();
    message.leaderSteamId = object.leaderSteamId ?? "0";
    message.members = object.members?.map((e) => AppTeamInfo_Member.fromPartial(e)) || [];
    message.mapNotes = object.mapNotes?.map((e) => AppTeamInfo_Note.fromPartial(e)) || [];
    message.leaderMapNotes = object.leaderMapNotes?.map((e) => AppTeamInfo_Note.fromPartial(e)) || [];
    return message;
  },
};

function createBaseAppTeamInfo_Member(): AppTeamInfo_Member {
  return { steamId: "0", name: "", x: 0, y: 0, isOnline: false, spawnTime: 0, isAlive: false, deathTime: 0 };
}

export const AppTeamInfo_Member: MessageFns<AppTeamInfo_Member> = {
  encode(message: AppTeamInfo_Member, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.steamId !== "0") {
      writer.uint32(8).uint64(message.steamId);
    }
    if (message.name !== "") {
      writer.uint32(18).string(message.name);
    }
    if (message.x !== 0) {
      writer.uint32(29).float(message.x);
    }
    if (message.y !== 0) {
      writer.uint32(37).float(message.y);
    }
    if (message.isOnline !== false) {
      writer.uint32(40).bool(message.isOnline);
    }
    if (message.spawnTime !== 0) {
      writer.uint32(48).uint32(message.spawnTime);
    }
    if (message.isAlive !== false) {
      writer.uint32(56).bool(message.isAlive);
    }
    if (message.deathTime !== 0) {
      writer.uint32(64).uint32(message.deathTime);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AppTeamInfo_Member {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAppTeamInfo_Member();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.steamId = reader.uint64().toString();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 29) {
            break;
          }

          message.x = reader.float();
          continue;
        }
        case 4: {
          if (tag !== 37) {
            break;
          }

          message.y = reader.float();
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.isOnline = reader.bool();
          continue;
        }
        case 6: {
          if (tag !== 48) {
            break;
          }

          message.spawnTime = reader.uint32();
          continue;
        }
        case 7: {
          if (tag !== 56) {
            break;
          }

          message.isAlive = reader.bool();
          continue;
        }
        case 8: {
          if (tag !== 64) {
            break;
          }

          message.deathTime = reader.uint32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<AppTeamInfo_Member>, I>>(base?: I): AppTeamInfo_Member {
    return AppTeamInfo_Member.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AppTeamInfo_Member>, I>>(object: I): AppTeamInfo_Member {
    const message = createBaseAppTeamInfo_Member();
    message.steamId = object.steamId ?? "0";
    message.name = object.name ?? "";
    message.x = object.x ?? 0;
    message.y = object.y ?? 0;
    message.isOnline = object.isOnline ?? false;
    message.spawnTime = object.spawnTime ?? 0;
    message.isAlive = object.isAlive ?? false;
    message.deathTime = object.deathTime ?? 0;
    return message;
  },
};

function createBaseAppTeamInfo_Note(): AppTeamInfo_Note {
  return { type: 0, x: 0, y: 0 };
}

export const AppTeamInfo_Note: MessageFns<AppTeamInfo_Note> = {
  encode(message: AppTeamInfo_Note, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.type !== 0) {
      writer.uint32(16).int32(message.type);
    }
    if (message.x !== 0) {
      writer.uint32(29).float(message.x);
    }
    if (message.y !== 0) {
      writer.uint32(37).float(message.y);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AppTeamInfo_Note {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAppTeamInfo_Note();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.type = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 29) {
            break;
          }

          message.x = reader.float();
          continue;
        }
        case 4: {
          if (tag !== 37) {
            break;
          }

          message.y = reader.float();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<AppTeamInfo_Note>, I>>(base?: I): AppTeamInfo_Note {
    return AppTeamInfo_Note.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AppTeamInfo_Note>, I>>(object: I): AppTeamInfo_Note {
    const message = createBaseAppTeamInfo_Note();
    message.type = object.type ?? 0;
    message.x = object.x ?? 0;
    message.y = object.y ?? 0;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}